(svn r19439) [1.0] -Backport from trunk: - Fix: Custom group names are misaligned with default ones when using rtl languages [FS#3700] (r19438) - Fix: With certain game settings one could clear tiles for free when building long roads (r19436) - Fix: When loading a savegame created with a house NewGRF without that NewGRF available all houses became tall office blocks (r19435) - Fix: Limit rail clearance earnings to 3/4s of rail build cost, to avoid money making loophole when rail build cost is less than rail removal earnings (r19433) - Fix: Crash when the error message 'owned by <town>' was shown [FS#3696] (r19432) - Feature: Append rail type speed limit (if set) to rail type selection list, and toolbar title (r19431)

+++ src/newgrf_commons.h
@@ -57,7 +57,7 @@ public:
 	void Add(uint8 local_id, uint32 grfid, uint entity_type);
 	virtual uint16 AddEntityID(byte grf_local_id, uint32 grfid, byte substitute_id);
 
-	uint16 GetSubstituteID(byte entity_id);
+	uint16 GetSubstituteID(uint16 entity_id);
 	virtual uint16 GetID(uint8 grf_local_id, uint32 grfid);
 
 	inline uint16 GetMaxMapping() { return max_new_entities; }
